# Admin/bin/Debug/netcoreapp3.1/Docs/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Admin API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/users/ours": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetUsers",
        "parameters": [
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortKey13"
            },
            "x-position": 1
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Order"
            },
            "x-position": 2
          },
          {
            "name": "DomainId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "UserGroupId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "AccountName",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "Name",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "StartDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 7
          },
          {
            "name": "StartDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 8
          },
          {
            "name": "EndDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 9
          },
          {
            "name": "EndDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 10
          },
          {
            "name": "IncludingEndDateNotSet",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "x-position": 11
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 12
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000.0,
              "minimum": 1.0
            },
            "x-position": 13
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      }
    },
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetUsers2",
        "parameters": [
          {
            "name": "OrganizationCode",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortKey13"
            },
            "x-position": 2
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Order"
            },
            "x-position": 3
          },
          {
            "name": "DomainId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "UserGroupId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "AccountName",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "Name",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 7
          },
          {
            "name": "StartDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 8
          },
          {
            "name": "StartDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 9
          },
          {
            "name": "EndDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 10
          },
          {
            "name": "EndDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 11
          },
          {
            "name": "IncludingEndDateNotSet",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "x-position": 12
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 13
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000.0,
              "minimum": 1.0
            },
            "x-position": 14
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      }
    },
    "/api/users/ours/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetUser2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_PutUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "param",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutUserParam"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_PutUser2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "param",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutUserParam"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Organization": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "address",
          "phone",
          "url",
          "adminPhone",
          "adminMail",
          "distributionServerIp",
          "startDate",
          "isValid"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "address": {
            "type": "string",
            "minLength": 1
          },
          "phone": {
            "type": "string",
            "format": "phone",
            "maxLength": 11,
            "minLength": 10
          },
          "url": {
            "type": "string",
            "format": "uri",
            "minLength": 1
          },
          "adminPhone": {
            "type": "string",
            "format": "phone",
            "maxLength": 11,
            "minLength": 10
          },
          "adminMail": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "distributionServerIp": {
            "type": "string",
            "minLength": 1,
            "pattern": "((2(5[0-5]|[0-4]\\d))|[0-1]?\\d{1,2})(\\.((2(5[0-5]|[0-4]\\d))|[0-1]?\\d{1,2})){3}"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isValid": {
            "type": "boolean"
          }
        }
      },
...
      "Domain": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "organizationCode": {
            "type": "integer",
            "format": "int32"
          },
          "organization": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Organization"
              }
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "adObjectId": {
            "type": "string",
            "format": "guid"
          }
        }
      },
...
      "EndUser": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "x-abstract": true,
            "additionalProperties": false,
            "required": [
              "authenticateWhenUnlockingScreen"
            ],
            "properties": {
              "domainId": {
                "type": "string",
                "format": "guid"
              },
              "domain": {
                "$ref": "#/components/schemas/Domain"
              },
              "userGroupEndUsers": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UserGroupEndUser"
                }
              },
              "authenticateWhenUnlockingScreen": {
                "type": "boolean"
              },
              "availablePeriods": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AvailablePeriod"
                }
              },
              "adObjectId": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
      "UserGroupEndUser": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "userGroupId": {
            "type": "string",
            "format": "guid"
          },
          "userGroup": {
            "$ref": "#/components/schemas/UserGroup"
          },
          "endUserId": {
            "type": "string",
            "format": "guid"
          },
          "endUser": {
            "$ref": "#/components/schemas/EndUser"
          }
        }
      },
      "UserGroup": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "domainId": {
            "type": "string",
            "format": "guid"
          },
          "domain": {
            "$ref": "#/components/schemas/Domain"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "userGroupEndUsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserGroupEndUser"
            }
          },
          "adObjectId": {
            "type": "string",
            "format": "guid"
          }
        }
      },
      "AvailablePeriod": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "startDate"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "endUserId": {
            "type": "string",
            "format": "guid"
          },
          "endUser": {
            "$ref": "#/components/schemas/EndUser"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "User": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false,
        "required": [
          "accountName"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "accountName": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "Order": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Asc",
          "Desc"
        ],
        "enum": [
          0,
          1
        ]
      },
...
      "PaginatedResponseOfUserDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "results": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          }
        }
      },
      "UserDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "accountName": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "domainId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "authenticateWhenUnlockingScreen": {
            "type": "boolean",
            "nullable": true
          },
          "domain": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Domain"
              }
            ]
          },
          "userGroups": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/UserGroup"
            }
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "SortKey13": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "DomainName",
          "AccountName",
          "Name",
          "StartDate",
          "EndDate"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4
        ]
      },
      "PutUserParam": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "authenticateWhenUnlockingScreen"
        ],
        "properties": {
          "authenticateWhenUnlockingScreen": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ],
...
}